Ensure the same user won't send more than once the same request
For provider like AMDA (using regular Provider Cache decorator) Speasy
will ensure that each request is only sent once. Each identical request
will wait for the completion of the first one.

This will be specially good on the proxy for perfectly identical requests.
The main limitation here is that several overlapping requests can be sent
in parallel.
